+++ src/main/groovy/com/powsybl/dynawo/automatons/CurrentLimitAutomatonGroovyExtension.groovy
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawo.dynamicmodels
+package com.powsybl.dynawo.automatons
 
 import java.util.function.Consumer
