+++ src/test/java/com/powsybl/dynawo/DynawoGroovyDynamicModelsSupplierTest.java
@@ -14,8 +14,8 @@ import com.powsybl.dynamicsimulation.groovy.DynamicModelGroovyExtension;
 import com.powsybl.dynamicsimulation.groovy.GroovyDynamicModelsSupplier;
 import com.powsybl.dynamicsimulation.groovy.GroovyExtension;
 import com.powsybl.dynawo.automatons.CurrentLimitAutomaton;
+import com.powsybl.dynawo.automatons.CurrentLimitAutomatonGroovyExtension;
 import com.powsybl.dynawo.dynamicmodels.AbstractBlackBoxModel;
-import com.powsybl.dynawo.dynamicmodels.CurrentLimitAutomatonGroovyExtension;
 import com.powsybl.dynawo.dynamicmodels.GeneratorSynchronousFourWindingsProportionalRegulations;
 import com.powsybl.dynawo.dynamicmodels.LoadAlphaBeta;
 import com.powsybl.dynawo.dynamicmodels.OmegaRef;
